+++ agency-20250227/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-03-28 Version: 1.0.1
+- Update API GetSubPartnerList: add response parameters Body.SubPartnerList.
+- Update API GetSubPartnerList: delete response parameters Body.Data.
+
+
 2025-03-05 Version: 1.0.0
 - Generated python 2025-02-27 for Agency.